requirements.txt
streamlit==1.24.1 
pillow==9.4.0
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1